# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CMSFaqPageSuccessResponse200:
    properties:
      items:
        $ref: '#/definitions/dto.FaqPageResponse'
      message:
        example: Faq page retrieved successfully
        type: string
    type: object
  dto.CMSFaqPagesSuccessResponse200:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.FaqPageResponse'
        type: array
      limit:
        example: 10
        type: integer
      message:
        example: Faq page retrieved successfully
        type: string
      page:
        example: 1
        type: integer
      totalCount:
        example: 100
        type: integer
    type: object
...
  dto.CreateFaqContentRequest:
    properties:
      authored_at:
        type: string
      authored_on:
        type: string
      categories:
        items:
          $ref: '#/definitions/dto.CreateCategoryId'
        type: array
      components:
        items:
          $ref: '#/definitions/dto.CreateComponentRequest'
        type: array
      html_input:
        type: string
      language:
        type: string
      meta_tag:
        $ref: '#/definitions/dto.CreateMetaTagRequest'
      mode:
        type: string
      publish_on:
        type: string
      publish_status:
        type: string
      revision:
        $ref: '#/definitions/dto.CreateRevisionRequest'
      title:
        type: string
      unpublish_on:
        type: string
      url:
        type: string
      url_alias:
        type: string
      workflow_status:
        type: string
    type: object
  dto.CreateFaqPageRequest:
    properties:
      contents:
        items:
          $ref: '#/definitions/dto.CreateFaqContentRequest'
        type: array
    type: object
...
  dto.ErrorResponse400:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Bad Request
        type: string
    type: object
...
  dto.ErrorResponse500:
    properties:
      error:
        example: Internal Server Error
        type: string
      message:
        example: An unexpected error occurred on the server
        type: string
    type: object
...
info:
  contact: {}
  description: This API supports both Web and admin (CMS) operations, including authentication,
    page management.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /cms/faqpages:
    get:
      description: Retrieve a list of FAQ pages with optional filtering, sorting,
        pagination, and language selection.
      parameters:
      - description: Filter criteria (URL-encoded JSON string), e.g., category_faq
          and category_keywords
        in: query
        name: query
        type: string
      - description: Sorting fields, e.g., 'title:asc,updated_at:desc'
        in: query
        name: sort
        type: string
      - description: Page number for pagination (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default is 10)
        in: query
        name: limit
        type: integer
      - description: Language code for localized content (e.g., en, th)
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqPagesSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: List FAQ Pages
      tags:
      - CMS - Faq Pages
    post:
      consumes:
      - application/json
      description: |-
        Create a new FAQ page with optional components such as categories, components, and metatag.
        The request must include exactly one content block.
        A valid revision must be included in the request other parts are optional.
      parameters:
      - description: 'FAQ Page payload (optional: categories, components)'
        in: body
        name: faq_page_data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFaqPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqPageSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Create a New FAQ Page
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{contentId}/contents:
    put:
      consumes:
      - application/json
      description: |-
        Update an existing FAQ content by its content ID.
        This operation does not merge with previous content — it overwrites everything.
        A valid revision must be included in the request.
      parameters:
      - description: FAQ Content ID (UUID)
        in: path
        name: contentId
        required: true
        type: string
      - description: Updated FAQ Content
        in: body
        name: faqContent
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFaqContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqContentSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Update FAQ Content
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{pageId}:
    delete:
      description: Delete an existing FAQ page using its unique identifier.
      parameters:
      - description: FAQ Page ID
        in: path
        name: pageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Delete FAQ Page
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{pageId}/contents/{languageCode}:
    delete:
      description: Delete the content of a specific FAQ page using its unique identifier,
        language code, and mode.
      parameters:
      - description: FAQ Page ID (UUID)
        in: path
        name: pageId
        required: true
        type: string
      - description: Language Code (e.g., en, th)
        in: path
        name: languageCode
        required: true
        type: string
      - description: Mode (e.g., draft, published, histories, preview). Defaults to
          'published'.
        in: query
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Delete FAQ Page Content
      tags:
      - CMS - Faq Pages
    get:
      description: Retrieve the content of a specific FAQ page by its unique identifier
        and language code.
      parameters:
      - description: FAQ Page ID
        in: path
        name: pageId
        required: true
        type: string
      - description: Language Code (e.g., en, th)
        in: path
        name: languageCode
        required: true
        type: string
      - description: Mode (e.g., draft, published, histories, preview). Defaults to
          'published'.
        in: query
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqContentSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Get FAQ Page Content
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{pageId}/latestcontent/{languageCode}:
    get:
      description: Retrieve the latest content for FAQ page.
      parameters:
      - description: FAQ Page ID (UUID format)
        in: path
        name: pageId
        required: true
        type: string
      - description: Language Code (e.g., en, th)
        in: path
        name: languageCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqContentSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Get Latest FAQ Page Content
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{pageid}:
    get:
      description: Retrieve a specific FAQ page using its unique identifier.
      parameters:
      - description: FAQ Page ID
        in: path
        name: pageid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqPageSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Get FAQ Page by ID
      tags:
      - CMS - Faq Pages
  /cms/faqpages/{revisionId}/revisions:
    post:
      consumes:
      - application/json
      description: Revert FAQ content using a specific revision ID.
      parameters:
      - description: Revision ID (UUID)
        in: path
        name: revisionId
        required: true
        type: string
      - description: Revision payload
        in: body
        name: revision
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRevisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqContentSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Revert FAQ Page Content
      tags:
      - CMS - Faq Pages
  /cms/faqpages/category/{categoryTypeCode}/{pageId}/{languageCode}:
    get:
      description: Retrieve the categories of a specific FAQ page using its unique
        identifier, category type code, language code, and mode.
      parameters:
      - description: FAQ Page ID (UUID)
        in: path
        name: pageId
        required: true
        type: string
      - description: Category Type Code
        in: path
        name: categoryTypeCode
        required: true
        type: string
      - description: Language Code (e.g., en, th)
        in: path
        name: languageCode
        required: true
        type: string
      - description: Mode (e.g., draft, published, histories, preview). Defaults to
          'published'.
        in: query
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqCategoriesSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Get FAQ Page Categories
      tags:
      - CMS - Faq Pages
  /cms/faqpages/duplicate/{contentId}/contents:
    post:
      consumes:
      - application/json
      description: Duplicate an existing FAQ content using its unique identifier and
      parameters:
      - description: FAQ content ID (UUID)
        in: path
        name: contentId
        required: true
        type: string
      - description: Revision payload
        in: body
        name: revision
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRevisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CMSFaqContentSuccessResponse200'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse500'
      summary: Duplicate FAQ Content to another language
      tags:
      - CMS - Faq Pages
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"